requirements.txt
streamlit
st-gsheets-connection
pandas
openai
python-dotenv
koreanize-matplotlib
seaborn